requirements.txt
numpy>=1.26
matplotlib>=3.8
scikit-learn>=1.5
pillow>=10.0
opencv-python>=4.10
torch>=2.2
torchvision>=0.17
tqdm>=4.66
seaborn>=0.13
jupyter>=1.0
